requirements.txt
kivy
opencv-python
filestack-python